requirements.txt
Click
Envelopes
openpyxl
pyinstaller
pymssql
python-dotenv
workdays
